requirements.txt
langchain
langchain-community
langchain-google-genai
# langchain-experimental
# PyMySQL
# sentence-transformers
chromadb
# langchain_huggingface
# langchain-cohere
streamlit
python-dotenv